requirements.txt
tkinter
redis
python-dotenv
import bcrypt  